split into binary and client packages
