+++ rust/forevervm-client/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "forevervm-client"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+chrono = { version = "0.4.39", features = ["serde"] }
+futures-util = "0.3.31"
+reqwest = { version = "0.12.12", features = ["json"] }
+rustls = "0.23.21"
+serde = { version = "1.0.217", features = ["derive"] }
+serde_json = "1.0.137"
+thiserror = "2.0.11"
+tokio = "1.43.0"
+tokio-tungstenite = { version = "0.26.1", features = ["rustls-tls-webpki-roots"] }
+tracing = "0.1.41"
+tungstenite = "0.26.1"
+url = "2.5.4"
