+++ rust/forevervm-client/src/client/error.rs
