+++ rust/forevervm-client/src/client/util.rs
