+++ rust/forevervm-client/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod api;
+pub mod client;
