+++ rust/forevervm/Cargo.toml
@@ -9,16 +9,17 @@ chrono = { version = "0.4.39", features = ["serde"] }
 clap = { version = "4.4", features = ["derive", "env"] }
 colorize = "0.1.0"
 dirs = "6.0.0"
-futures-util = "0.3.31"
-reqwest = { version = "0.12.12", features = ["json"] }
+# futures-util = "0.3.31"
+forevervm-client = { path = "../forevervm-client" }
+# reqwest = { version = "0.12.12", features = ["json"] }
 rpassword = "7.3.1"
-rustls = "0.23.21"
+# rustls = "0.23.21"
 rustyline = "15.0.0"
 serde = { version = "1.0.217", features = ["derive"] }
 serde_json = "1.0.137"
-thiserror = "2.0.11"
+# thiserror = "2.0.11"
 tokio = { version = "1.43.0", features = ["macros", "rt-multi-thread"] }
-tokio-tungstenite = { version = "0.26.1", features = ["rustls-tls-webpki-roots"] }
-tracing = "0.1.41"
-tungstenite = "0.26.1"
+# tokio-tungstenite = { version = "0.26.1", features = ["rustls-tls-webpki-roots"] }
+# tracing = "0.1.41"
+# tungstenite = "0.26.1"
 url = { version = "2.5.4", features = ["serde"] }
