+++ rust/forevervm/src/commands/auth.rs
@@ -1,6 +1,8 @@
-use crate::{api::token::ApiToken, client::ForeverVMClient, config::ConfigManager};
 use colorize::AnsiColor;
+use forevervm_client::{api::token::ApiToken, client::ForeverVMClient};
 use url::Url;
+
+use crate::config::ConfigManager;
 
 pub async fn whoami() -> anyhow::Result<()> {
     let client = ConfigManager::new()?.client()?;
