+++ rust/forevervm/src/commands/repl.rs
@@ -1,8 +1,6 @@
-use crate::{
-    api::{api_types::ExecResultType, id_types::MachineName},
-    config::ConfigManager,
-};
+use crate::config::ConfigManager;
 use colorize::AnsiColor;
+use forevervm_client::api::{api_types::ExecResultType, id_types::MachineName};
 use rustyline::{error::ReadlineError, DefaultEditor};
 
 pub async fn machine_repl(machine_name: Option<MachineName>) -> anyhow::Result<()> {
