+++ rust/forevervm/src/lib.rs
@@ -1,8 +1,5 @@
-pub mod api;
-pub mod client;
 pub mod commands;
 pub mod config;
-pub mod error;
 pub mod util;
 
 pub const DEFAULT_SERVER_URL: &str = "https://api.forevervm.com";
